Move tests in the right position
